requirements.txt
pip == 20.3.1
pymongo==3.11.2
pytest==6.1.2
dnspython==2.0.0